feat(WIP): add a configure node modal
