+++ package.json
@@ -55,6 +55,7 @@
     "@types/node": "^20.3.1",
     "@types/react": "^18.2.13",
     "@types/react-dom": "^18.2.6",
+    "@types/urijs": "^1.19.19",
     "@types/uuid": "^9.0.2",
     "@typescript-eslint/eslint-plugin": "^5.60.0",
     "@typescript-eslint/parser": "^5.60.0",
@@ -89,6 +90,7 @@
     "react-router": "^6.13.0",
     "react-router-dom": "^6.13.0",
     "typescript": "^5.1.3",
+    "urijs": "^1.19.11",
     "vite": "^4.3.9",
     "zod": "^3.21.4"
   }
